Move database dumping to data module
